+++ lib/rubocop/cop/norb_cops.rb
@@ -4,7 +4,7 @@ require_relative 'mixin/target_rails_version'
 
 require_relative 'norb/active_record_namespaced'
 require_relative 'norb/active_record_through_business'
-require_relative 'norb/misplaced_branching_logic'
+require_relative 'norb/branching_logic'
 require_relative 'norb/misplaced_logic'
 require_relative 'norb/one_controller_action_instance_variable'
 require_relative 'norb/standard_restful_controller_actions'