+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### New features
 
-* Adding `Norb/MisplacedBranchingLogic` cop to ensure branching logic is not in the wrong place. ([@tleish][])
+* Adding `Norb/BranchingLogic` cop to ensure branching logic is not in the wrong place. ([@tleish][])
 * Adding `Norb/ActiveRecordNamespaced` cop to ensure ActiveRecord models are namespaced. ([@tleish][])
 * Adding `Norb/ActiveRecordThroughBusiness` cop to ensure direct ActiveRecord calls come from business objects. ([@tleish][])
 * Adding `Norb/StandardRestfulControllerActions` cop to ensure code logic is not in the wrong place. ([@tleish][])